refactor: fixed sonarcloud issues in server component
